# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmokeWatch Simulator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SimulationInput {
  /** Twitch or Kick stream URL */
  streamUrl: string;
  /** Minimum minutes between detections */
  cooldownMinutes?: number;
  /** Minimum confidence score to trigger (0–1) */
  confidenceThreshold?: number;
}

export type SimulationSessionStatus = typeof SimulationSessionStatus[keyof typeof SimulationSessionStatus];


export const SimulationSessionStatus = {
  running: 'running',
  stopped: 'stopped',
} as const;

export interface SimulationSession {
  sessionId: string;
  streamUrl: string;
  status: SimulationSessionStatus;
  startedAt: string;
  /** @nullable */
  stoppedAt?: string | null;
  detectionCount: number;
  cooldownMinutes: number;
  confidenceThreshold: number;
  /** @nullable */
  lastDetectedAt?: string | null;
}

export interface DetectionEvent {
  eventId: string;
  sessionId: string;
  streamUrl: string;
  timestamp: string;
  confidence: number;
  /** @nullable */
  messageSent: string | null;
  blockedByCooldown: boolean;
  frameNumber?: number;
}

export interface MessagePool {
  messages: string[];
  link: string;
}