# lib/api-zod/src/generated/types/detectionEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmokeWatch Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface DetectionEvent {
  eventId: string;
  sessionId: string;
  streamUrl: string;
  timestamp: string;
  confidence: number;
  /** @nullable */
  messageSent: string | null;
  blockedByCooldown: boolean;
  frameNumber?: number;
}