# lib/api-zod/src/generated/types/messagePool.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmokeWatch Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface MessagePool {
  messages: string[];
  link: string;
}