# lib/api-zod/src/generated/types/simulationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmokeWatch Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface SimulationInput {
  /** Twitch or Kick stream URL */
  streamUrl: string;
  /** Minimum minutes between detections */
  cooldownMinutes?: number;
  /** Minimum confidence score to trigger (0–1) */
  confidenceThreshold?: number;
}