# lib/api-zod/src/generated/types/simulationSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmokeWatch Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { SimulationSessionStatus } from './simulationSessionStatus';

export interface SimulationSession {
  sessionId: string;
  streamUrl: string;
  status: SimulationSessionStatus;
  startedAt: string;
  /** @nullable */
  stoppedAt?: string | null;
  detectionCount: number;
  cooldownMinutes: number;
  confidenceThreshold: number;
  /** @nullable */
  lastDetectedAt?: string | null;
}